% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBollenkaart.R
\name{plot.bollenkaart}
\alias{plot.bollenkaart}
\title{In dit script kan met een data.frame met afwijkingen en een rasterlayer objext een bollenkaart gemaakt worden}
\usage{
\method{plot}{bollenkaart}(title, data, diffcol, xcol, ycol, laycol,
  minlay, maxlay, outfolder, outname, shapename, shapefolder, GGRaster,
  xmin, xmax, ymin, ymax, width, height)
}
\arguments{
\item{title}{text die boven de plot komt te staan}

\item{data}{data frame met afwijkingen die geplot moeten worden}

\item{diffcol}{kolom in data waarin de afwijkingen per laag staan}

\item{xcol}{kolom in data waar de x-waarde van het meetpunt staat}

\item{ycol}{komom in data waar de y-waarde van het meetpunt staat}

\item{minlay}{bovenste laag die geplot moet worden, gebruiken voor meerdere modellagen binnen een wvp}

\item{maxlay}{onderste laag die geplot moet worden, gebruiken voor meerdere modellagen binnen een wvp}

\item{outfolder}{map waarin de .png file moet worden opgeslagen}

\item{outname}{naam van de .png}

\item{shapename}{naam van een shapefile die ter orientatie over het figuur geplot wordt (zonder .shp)}

\item{shapefolder}{map waarin de shapefile staat}

\item{GGRaster}{Rasterlayer object dat geplot moet worden}

\item{xmin}{minimale en maximale x- en y-waardes van het plotwindow}

\item{xmax}{minimale en maximale x- en y-waardes van het plotwindow}

\item{ymin}{minimale en maximale x- en y-waardes van het plotwindow}

\item{ymax}{minimale en maximale x- en y-waardes van het plotwindow}

\item{width}{breedte van de png in cm}

\item{height}{hoogte van de png in cm}

\item{laycolkolom}{in data waar de modellaag staat}
}
\value{
.png met bollenkaart
}
\description{
In dit script kan met een data.frame met afwijkingen en een rasterlayer objext een bollenkaart gemaakt worden
}
