% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotdHead.R
\name{plot.dhead}
\alias{plot.dhead}
\title{Dit is de functie waarmee de rasterlayerobject wordt geplot in een .png bestand}
\usage{
\method{plot}{dhead}(title, outfolder, outname, shapename, shapefolder,
  dHeadRaster, xmin, xmax, ymin, ymax, width, height)
}
\arguments{
\item{title}{text die boven de plot komt te staan}

\item{outfolder}{map waarin de .png file moet worden opgeslagen}

\item{outname}{naam van de .png}

\item{shapename}{naam van een shapefile die ter orientatie over het figuur geplot wordt (zonder .shp)}

\item{shapefolder}{map waarin de shapefile staat}

\item{dHeadRaster}{Rasterlayer object dat geplot moet worden}

\item{xmin}{minimale en maximale x- en y-waardes van het plotwindow}

\item{xmax}{minimale en maximale x- en y-waardes van het plotwindow}

\item{ymin}{minimale en maximale x- en y-waardes van het plotwindow}

\item{ymax}{minimale en maximale x- en y-waardes van het plotwindow}

\item{width}{breedte van de png in cm}

\item{height}{hoogte van de png in cm}
}
\description{
Dit is de functie waarmee de rasterlayerobject wordt geplot in een .png bestand
}
