% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTijdreeksen.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Deze functies maakt tijdreeksen van iMODFLOW bestanden.}
\usage{
\method{plot}{timeseries}(ts_folder, outfolder, metafile, namecol, laycol,
  nlayers, datemin, datemax, width = 16, height = 24.7/3)
}
\arguments{
\item{ts_folder}{map waarin .txt bestanden staat met daarin gemeten en berekende grondwaterstanden. Dit wordt aangemaakt met de idf timeserietool}

\item{outfolder}{locatie waar alle tijdreeksen weggeschreven worden. voor elke modellaag wordt een map aangemaakt}

\item{metafile}{File met daarin minimaal de volgende data: naamkolom en kolom met de modellaag van de peilbuis. Dit kan zowel een csv als een ipf zijn}

\item{namecol}{kolom met de naam van de peilbuis. Deze moet overeenkomen met de naam van de .txt bestanden.}

\item{laycol}{kolom met de modellaag van de peilbuisfilter}

\item{nlayers}{aantal lagen dat het model heeft}

\item{datemin}{minimum datum grafiek}

\item{datemax}{maximum datum grafiek}

\item{width}{breedte grafiek in cm, standaard 16 cm}

\item{height}{hoogte grafiek in cm, standaard 24,7/3}
}
\description{
Deze functies maakt tijdreeksen van iMODFLOW bestanden.
}
