% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadHydroconnect.R
\name{read.hydroconnect}
\alias{read.hydroconnect}
\title{In dit script staat de functie om csv-files in het hydroconnect format te lezen.}
\usage{
read.hydroconnect(csv)
}
\arguments{
\item{csv}{naam .csv bestand met extentie}
}
\value{
output van deze functie is een lijst met de metadata en de data
}
\description{
In dit script staat de functie om csv-files in het hydroconnect format te lezen.
}
