% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPFreadwrite.R
\name{read.ipf}
\alias{read.ipf}
\title{functie om ipf-files in te lezen.}
\usage{
read.ipf(ipf_file, sep = ",")
}
\arguments{
\item{ipf_file}{ipf file input}

\item{sep}{separator in the ipf file. By default sep is ","}
}
\value{
returns a dataframe with the ipf data
}
\description{
functie om ipf-files in te lezen.
}
