% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2theme.R
\name{theme1}
\alias{theme1}
\title{Thema om tijdreeksen te plotten}
\usage{
theme1(base_size = 10, base_family = "Arial")
}
\arguments{
\item{base_size}{Grootte van het lettertype op de plot. Default: 10}

\item{base_family}{Lettertype van de plot. Default: Arial}
}
\value{
Output opmaak van de plot
}
\description{
Thema om tijdreeksen te plotten
}
