% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPFreadwrite.R
\name{write.ipf}
\alias{write.ipf}
\title{functie om ipf-files weg te schrijven.}
\usage{
write.ipf(x, ipf_out = "output.ipf", txtcol = 0, sep = ",")
}
\arguments{
\item{x}{dataframe to convert to ipf file}

\item{ipf_out}{name of the output ipf, including extension. By default "output.ipf" is generated}

\item{txtcol}{column in which the ipf file refers to an ipf text file. By default this parameter is 0 (no text file)}

\item{sep}{separator in the ipf file. By default sep is ","}
}
\value{
returns an ipf file
}
\description{
functie om ipf-files weg te schrijven.
}
